Improve closeDevice function
The previous closeDevice function in the Kinect.sc class did not account for the deviceCount. Additionally, the current device which was closed was the last opened device by default. However, the user would have no way to specify closing another device which was opened in the past. For this reason, the mOpenDevices unordered map was added to kinectData in Kinect.cpp to help keep track of a dictionary of open devices. This makes opening and closing devices efficient (constant time amortized) and effective (whereas this feature previously did not work as intended).